Remove lucene label scan index
- Deleting all classes related to lucene implementation of
  label scan store.
- Collaps label scan store tests that previously had one
  implementation for lucene and one for native.
- Remove LUCENE from LabelIndex enum in settings.
- Fix compilation errors.
