--- enterprise/ha/src/test/java/org/neo4j/kernel/api/impl/labelscan/LuceneLabelScanStoreHaIT.java
@@ -1,38 +1,0 @@
-/*
- * Copyright (c) 2002-2017 "Neo Technology,"
- * Network Engine for Objects in Lund AB [http://neotechnology.com]
- *
- * This file is part of Neo4j.
- *
- * Neo4j is free software: you can redistribute it and/or modify
- * it under the terms of the GNU Affero General Public License as
- * published by the Free Software Foundation, either version 3 of the
- * License, or (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU Affero General Public License for more details.
- *
- * You should have received a copy of the GNU Affero General Public License
- * along with this program. If not, see <http://www.gnu.org/licenses/>.
- */
-package org.neo4j.kernel.api.impl.labelscan;
-
-import org.neo4j.graphdb.factory.GraphDatabaseSettings.LabelIndex;
-import org.neo4j.kernel.extension.KernelExtensionFactory;
-
-public class LuceneLabelScanStoreHaIT extends LabelScanStoreHaIT
-{
-    @Override
-    protected KernelExtensionFactory<?> labelScanStoreExtension()
-    {
-        return new LuceneLabelScanStoreExtension();
-    }
-
-    @Override
-    protected String labelIndexSettingName()
-    {
-        return LabelIndex.LUCENE.name();
-    }
-}
